api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.15.2
pydantic-settings==2.9.1
python-dotenv==1.1.0
pyjwt==2.10.1
bcrypt==4.3.0
python-multipart==0.0.20
httpx==0.28.1
